workers/video/requirements.txt
celery==5.2.3
numpy==1.22.3
pillow==9.0.1
requests
ffmpeg-python
redis==4.1.4